fix: resolve 26 maintainability violations (batch 2)
File splits (5 files over 300 lines):
- Extract _rate_limit.py from api/app.py (314→198 lines)
- Extract _provider_cache.py from analysis/subprocess.py (304→261)
- Extract _req_mapping.py from core/evidence/parser.py (345→287)
- Extract _git_scoring.py from analysis/subagents/priority.py (312→247)
- Extract _renderers.py from analysis/prompts/builder.py (308→225)

JSX component extractions (10):
- StandardCard: extract CardActions, EyeToggle sub-components
- ImportModal: extract PickStep/ImportingStep/ErrorStep/WarningsStep/ConflictStep
- StandardDetail: extract EmptyState sub-component
- LiveViolationsFeed: extract ViolationDetail sub-component
- ReferenceEditor: extract CweList, CweRefInputs, GenericRefInputs
- SettingsPage: extract persistSetting helper, move AI_CMD_STORAGE_KEY to constants
- HistoryPage: proper catch fallback instead of empty catch

Tests + minor (9):
- Add HTTP status assertion in test_power_selector_backend
- Move imports to module level in 3 test files
- Parametrize error class test
- Replace mutable globals with frozen _PenaltyConfig dataclass
- Add reset_priority_config_cache() for test isolation
- Deduplicate compiled_dir computation in _consolidated.py
- Fix patch paths for moved _run_git_log
