+++ src/quodeq/analysis/subagents/_consolidated.py
@@ -17,12 +17,12 @@ from quodeq.shared.logging import log_info, log_warning
 
 def _build_consolidated_config(
     config: "RunConfig", dimensions: list[str], files_per_agent: int,
+    compiled_dir: "Path | None" = None,
 ) -> AnalysisConfig:
     """Build AnalysisConfig for consolidated mode."""
     # Deferred import: breaks circular dependency between _consolidated and runner.
     from quodeq.analysis.subagents.runner import _default_subagent_model
 
-    compiled_dir = (config.standards_dir / "compiled") if config.standards_dir else None
     subagent_model = config.options.subagent_model or _default_subagent_model()
     pool_budget_val = config.options.pool_budget
     return AnalysisConfig(
@@ -40,6 +40,7 @@ def _build_consolidated_config(
 def _collect_consolidated_results(
     config: "RunConfig", dimensions: list[str], ctx: Any,
     results: list[Any], evidence_dir: Path,
+    compiled_dir: "Path | None" = None,
 ) -> dict[str, Evidence]:
     """Deduplicate and parse consolidated results into per-dimension Evidence."""
     from quodeq.analysis.stream.counters import count_files_in_stream
@@ -63,7 +64,6 @@ def _collect_consolidated_results(
         module=config.target.name if config.target else "",
     )
 
-    compiled_dir = (config.standards_dir / "compiled") if config.standards_dir else None
     return parse_jsonl_to_evidence_by_dimension(
         merged_jsonl, ev_ctx, compiled_dir=compiled_dir,
         evaluators_dir=config.evaluators_dir,
@@ -96,6 +96,7 @@ def process_consolidated_dimensions(
     """Run all dimensions in a single pass -- files read once, not per dimension."""
     from quodeq.analysis.subagents.runner import _list_source_files, _compute_files_per_agent
 
+    compiled_dir = (config.standards_dir / "compiled") if config.standards_dir else None
     evidence_dir = config.work_dir or config.src
 
     # 1. List source files
@@ -112,7 +113,7 @@ def process_consolidated_dimensions(
     log_info(f"Consolidated analysis: {len(files)} files, {len(dimensions)} dimensions, max {config.options.max_subagents} agents")
 
     # 3. Build config and launch pool
-    base_ac = _build_consolidated_config(config, dimensions, files_per_agent)
+    base_ac = _build_consolidated_config(config, dimensions, files_per_agent, compiled_dir=compiled_dir)
     pool = SubagentPool(
         paths=PoolPaths(work_dir=config.src, evidence_dir=evidence_dir, queue_path=queue_path),
         options=PoolOptions(
@@ -125,4 +126,4 @@ def process_consolidated_dimensions(
     results = pool.run()
 
     # 4. Collect and return per-dimension evidence
-    return _collect_consolidated_results(config, dimensions, ctx, results, evidence_dir)
+    return _collect_consolidated_results(config, dimensions, ctx, results, evidence_dir, compiled_dir=compiled_dir)
