+++ src/quodeq/ui/src/constants.js
@@ -6,6 +6,8 @@ export const DEFAULT_POOL_BUDGET = 600;
 export const SUBAGENTS_STORAGE_KEY = 'cc-max-subagents';
 export const POOL_BUDGET_STORAGE_KEY = 'cc-pool-budget';
 
+export const AI_CMD_STORAGE_KEY = 'cc-ai-cmd';
+
 export const VISIBLE_STANDARDS_STORAGE_KEY = 'quodeq-visible-standards';
 export const DEFAULT_VISIBLE_STANDARDS = [
   'security', 'reliability', 'maintainability', 'performance', 'usability', 'flexibility',
