+++ src/quodeq/ui/src/features/history/components/HistoryPage.jsx
@@ -37,7 +37,7 @@ export default function HistoryPage({ trend, selection, availableRuns, dimension
         const d = new Date(entry.dateISO);
         return d.toLocaleDateString('en-GB', { day: 'numeric', month: 'long', year: 'numeric' })
           + ' ' + d.toLocaleTimeString('en-GB', { hour: '2-digit', minute: '2-digit' });
-      } catch { /* fall through */ }
+      } catch { return entry.dateISO || ''; }
     }
     return entry?.dateLabel || currentOverviewRun;
   }, [trend, currentOverviewRun]);
