+++ src/quodeq/ui/src/features/standards/components/ImportModal.jsx
@@ -4,6 +4,85 @@ import { importStandard } from '../../../api/index.js';
 const MAX_FILE_SIZE = 1024 * 1024; // 1MB
 const STEP = { PICK: 'pick', IMPORTING: 'importing', ERROR: 'error', WARNINGS: 'warnings', CONFLICT: 'conflict' };
 
+function PickStep({ fileRef, onFile, onClose }) {
+  return (
+    <>
+      <h3 className="modal-title">Import Evaluator</h3>
+      <p className="modal-body">Select a <strong>.quodeq</strong> file to import.</p>
+      <input ref={fileRef} type="file" accept=".quodeq,.json" onChange={onFile} style={{ margin: '12px 0' }} />
+      <div className="modal-actions">
+        <button type="button" className="btn-secondary" onClick={onClose}>Cancel</button>
+      </div>
+    </>
+  );
+}
+
+function ImportingStep() {
+  return (
+    <>
+      <h3 className="modal-title">Importing...</h3>
+      <p className="modal-body">Validating and importing evaluator.</p>
+    </>
+  );
+}
+
+function ErrorStep({ error, onClose }) {
+  return (
+    <>
+      <h3 className="modal-title">Import Failed</h3>
+      <p className="modal-body modal-body--warning">{error}</p>
+      <div className="modal-actions">
+        <button type="button" className="btn-secondary" onClick={onClose}>Close</button>
+      </div>
+    </>
+  );
+}
+
+function WarningsStep({ warnings, onClose, onProceed }) {
+  return (
+    <>
+      <h3 className="modal-title">Security Warnings</h3>
+      <p className="modal-body modal-body--warning">
+        This evaluator contains text that may attempt to manipulate the AI during analysis:
+      </p>
+      <ul className="modal-body" style={{ fontSize: '0.85rem', maxHeight: 200, overflow: 'auto' }}>
+        {warnings.map((w, i) => <li key={i}>{w}</li>)}
+      </ul>
+      <div className="modal-actions">
+        <button type="button" className="btn-secondary" onClick={onClose}>Cancel</button>
+        <button type="button" className="btn-primary" onClick={onProceed}>Import Anyway</button>
+      </div>
+    </>
+  );
+}
+
+function ConflictStep({ parsedData, conflict, warnings, onClose, onImportAsCopy, onOverwrite }) {
+  return (
+    <>
+      <h3 className="modal-title">ID Already Exists</h3>
+      <p className="modal-body">
+        A standard with ID <strong>{parsedData?.id}</strong> already exists
+        {conflict?.name ? ` ("${conflict.name}")` : ''}.
+      </p>
+      {warnings.length > 0 && (
+        <>
+          <p className="modal-body modal-body--warning" style={{ fontSize: '0.85rem' }}>
+            Security warnings were also detected:
+          </p>
+          <ul className="modal-body" style={{ fontSize: '0.8rem', maxHeight: 120, overflow: 'auto' }}>
+            {warnings.map((w, i) => <li key={i}>{w}</li>)}
+          </ul>
+        </>
+      )}
+      <div className="modal-actions">
+        <button type="button" className="btn-secondary" onClick={onClose}>Cancel</button>
+        <button type="button" className="btn-secondary" onClick={onImportAsCopy}>Import as Copy</button>
+        <button type="button" className="btn-danger" onClick={onOverwrite}>Overwrite</button>
+      </div>
+    </>
+  );
+}
+
 export default function ImportModal({ onClose, onImported }) {
   const [step, setStep] = useState(STEP.PICK); // pick | warnings | conflict | importing | error
   const [error, setError] = useState(null);
@@ -82,81 +161,19 @@ export default function ImportModal({ onClose, onImported }) {
   return (
     <div className="modal-overlay" onClick={onClose}>
       <div className="modal-dialog" onClick={(e) => e.stopPropagation()}>
-        {step === STEP.PICK && (
-          <>
-            <h3 className="modal-title">Import Evaluator</h3>
-            <p className="modal-body">
-              Select a <strong>.quodeq</strong> file to import.
-            </p>
-            <input
-              ref={fileRef}
-              type="file"
-              accept=".quodeq,.json"
-              onChange={handleFile}
-              style={{ margin: '12px 0' }}
-            />
-            <div className="modal-actions">
-              <button type="button" className="btn-secondary" onClick={onClose}>Cancel</button>
-            </div>
-          </>
-        )}
-
-        {step === STEP.IMPORTING && (
-          <>
-            <h3 className="modal-title">Importing...</h3>
-            <p className="modal-body">Validating and importing evaluator.</p>
-          </>
-        )}
-
-        {step === STEP.ERROR && (
-          <>
-            <h3 className="modal-title">Import Failed</h3>
-            <p className="modal-body modal-body--warning">{error}</p>
-            <div className="modal-actions">
-              <button type="button" className="btn-secondary" onClick={onClose}>Close</button>
-            </div>
-          </>
-        )}
-
-        {step === STEP.WARNINGS && (
-          <>
-            <h3 className="modal-title">Security Warnings</h3>
-            <p className="modal-body modal-body--warning">
-              This evaluator contains text that may attempt to manipulate the AI during analysis:
-            </p>
-            <ul className="modal-body" style={{ fontSize: '0.85rem', maxHeight: 200, overflow: 'auto' }}>
-              {warnings.map((w, i) => <li key={i}>{w}</li>)}
-            </ul>
-            <div className="modal-actions">
-              <button type="button" className="btn-secondary" onClick={onClose}>Cancel</button>
-              <button type="button" className="btn-primary" onClick={handleProceedWithWarnings}>Import Anyway</button>
-            </div>
-          </>
-        )}
-
+        {step === STEP.PICK && <PickStep fileRef={fileRef} onFile={handleFile} onClose={onClose} />}
+        {step === STEP.IMPORTING && <ImportingStep />}
+        {step === STEP.ERROR && <ErrorStep error={error} onClose={onClose} />}
+        {step === STEP.WARNINGS && <WarningsStep warnings={warnings} onClose={onClose} onProceed={handleProceedWithWarnings} />}
         {step === STEP.CONFLICT && (
-          <>
-            <h3 className="modal-title">ID Already Exists</h3>
-            <p className="modal-body">
-              A standard with ID <strong>{parsedData?.id}</strong> already exists
-              {conflict?.name ? ` ("${conflict.name}")` : ''}.
-            </p>
-            {warnings.length > 0 && (
-              <>
-                <p className="modal-body modal-body--warning" style={{ fontSize: '0.85rem' }}>
-                  Security warnings were also detected:
-                </p>
-                <ul className="modal-body" style={{ fontSize: '0.8rem', maxHeight: 120, overflow: 'auto' }}>
-                  {warnings.map((w, i) => <li key={i}>{w}</li>)}
-                </ul>
-              </>
-            )}
-            <div className="modal-actions">
-              <button type="button" className="btn-secondary" onClick={onClose}>Cancel</button>
-              <button type="button" className="btn-secondary" onClick={handleImportAsCopy}>Import as Copy</button>
-              <button type="button" className="btn-danger" onClick={handleForceImport}>Overwrite</button>
-            </div>
-          </>
+          <ConflictStep
+            parsedData={parsedData}
+            conflict={conflict}
+            warnings={warnings}
+            onClose={onClose}
+            onImportAsCopy={handleImportAsCopy}
+            onOverwrite={handleForceImport}
+          />
         )}
       </div>
     </div>
