+++ src/quodeq/ui/src/features/standards/components/StandardCard.jsx
@@ -67,14 +67,68 @@ function DuplicateModal({ standardId, onConfirm, onCancel }) {
   );
 }
 
+function CardActions({ standard, isDeletable, onDuplicate, onDownload, onDelete }) {
+  return (
+    <div className="standard-card-actions" onClick={(e) => e.stopPropagation()}>
+      <button type="button" className="card-action-btn" onClick={onDuplicate} title="Duplicate">
+        <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+          <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
+          <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
+        </svg>
+      </button>
+      <button type="button" className="card-action-btn" onClick={onDownload} title="Download">
+        <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+          <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
+          <polyline points="7 10 12 15 17 10" />
+          <line x1="12" y1="15" x2="12" y2="3" />
+        </svg>
+      </button>
+      {isDeletable && (
+        <button type="button" className="card-action-btn card-action-btn--danger" onClick={onDelete} title="Delete">
+          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+            <polyline points="3 6 5 6 21 6" />
+            <path d="M19 6l-1 14a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2L5 6" />
+            <path d="M10 11v6M14 11v6" />
+            <path d="M9 6V4a1 1 0 0 1 1-1h4a1 1 0 0 1 1 1v2" />
+          </svg>
+        </button>
+      )}
+    </div>
+  );
+}
+
+function EyeToggle({ isVisible, standardId, onToggleVisibility }) {
+  return (
+    <button
+      type="button"
+      className={`eye-toggle-btn${isVisible ? ' eye-toggle-btn--on' : ''}`}
+      title={isVisible ? 'Visible on Overview — click to hide' : 'Hidden from Overview — click to show'}
+      onClick={(e) => { e.stopPropagation(); onToggleVisibility(standardId); }}
+    >
+      {isVisible ? (
+        <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+          <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
+          <circle cx="12" cy="12" r="3" />
+        </svg>
+      ) : (
+        <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+          <path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94" />
+          <path d="M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19" />
+          <path d="M14.12 14.12a3 3 0 1 1-4.24-4.24" />
+          <line x1="1" y1="1" x2="23" y2="23" />
+        </svg>
+      )}
+    </button>
+  );
+}
+
 export default function StandardCard({ standard, onEdit, onDelete, onDuplicate, isVisible, onToggleVisibility }) {
   const [showDeleteModal, setShowDeleteModal] = useState(false);
   const [showDuplicateModal, setShowDuplicateModal] = useState(false);
 
   const principleCount = standard.principleCount ?? standard.principles?.length ?? 0;
   const requirementCount = standard.requirementCount ?? (standard.principles || []).reduce(
-    (sum, p) => sum + (p.requirements?.length ?? 0),
-    0
+    (sum, p) => sum + (p.requirements?.length ?? 0), 0
   );
   const isDeletable = standard.type !== 'builtin' && standard.type !== 'quodeq';
   const visClass = isVisible ? 'standard-card--visible' : 'standard-card--hidden';
@@ -89,79 +143,21 @@ export default function StandardCard({ standard, onEdit, onDelete, onDuplicate, 
           {standard.managed && (
             <span className="standard-managed-badge" title="Managed standard — read-only">managed</span>
           )}
-          <button
-            type="button"
-            className={`eye-toggle-btn${isVisible ? ' eye-toggle-btn--on' : ''}`}
-            title={isVisible ? 'Visible on Overview — click to hide' : 'Hidden from Overview — click to show'}
-            onClick={(e) => { e.stopPropagation(); onToggleVisibility(standard.id); }}
-          >
-            {isVisible ? (
-              <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-                <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
-                <circle cx="12" cy="12" r="3" />
-              </svg>
-            ) : (
-              <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-                <path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94" />
-                <path d="M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19" />
-                <path d="M14.12 14.12a3 3 0 1 1-4.24-4.24" />
-                <line x1="1" y1="1" x2="23" y2="23" />
-              </svg>
-            )}
-          </button>
+          <EyeToggle isVisible={isVisible} standardId={standard.id} onToggleVisibility={onToggleVisibility} />
         </div>
-
         <h3 className="standard-card-name">{standard.name}</h3>
-        {standard.description && (
-          <p className="standard-card-description">{standard.description}</p>
-        )}
-
+        {standard.description && <p className="standard-card-description">{standard.description}</p>}
         <div className="standard-card-counts">
           <span>{principleCount} {principleCount === 1 ? 'principle' : 'principles'}</span>
           <span className="standard-card-counts-sep">·</span>
           <span>{requirementCount} {requirementCount === 1 ? 'requirement' : 'requirements'}</span>
         </div>
-
-        <div className="standard-card-actions" onClick={(e) => e.stopPropagation()}>
-          <button
-            type="button"
-            className="card-action-btn"
-            onClick={() => setShowDuplicateModal(true)}
-            title="Duplicate"
-          >
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
-              <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
-              <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
-            </svg>
-          </button>
-          <button
-            type="button"
-            className="card-action-btn"
-            onClick={() => downloadStandard(standard.id)}
-            title="Download"
-          >
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
-              <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
-              <polyline points="7 10 12 15 17 10" />
-              <line x1="12" y1="15" x2="12" y2="3" />
-            </svg>
-          </button>
-          {isDeletable && (
-            <button
-              type="button"
-              className="card-action-btn card-action-btn--danger"
-              onClick={() => setShowDeleteModal(true)}
-              title="Delete"
-            >
-              <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
-                <polyline points="3 6 5 6 21 6" />
-                <path d="M19 6l-1 14a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2L5 6" />
-                <path d="M10 11v6M14 11v6" />
-                <path d="M9 6V4a1 1 0 0 1 1-1h4a1 1 0 0 1 1 1v2" />
-              </svg>
-            </button>
-          )}
-        </div>
+        <CardActions
+          isDeletable={isDeletable}
+          onDuplicate={() => setShowDuplicateModal(true)}
+          onDownload={() => downloadStandard(standard.id)}
+          onDelete={() => setShowDeleteModal(true)}
+        />
       </div>
 
       {showDeleteModal && (
