+++ src/quodeq/ui/src/features/standards/components/StandardDetail.jsx
@@ -5,6 +5,20 @@ function slugify(text) {
   return text.toLowerCase().trim().replace(/[^a-z0-9]+/g, '-').replace(/^-|-$/g, '');
 }
 
+function EmptyState() {
+  return (
+    <div className="detail-empty-state">
+      <svg width="36" height="36" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+        <circle cx="12" cy="12" r="10" />
+        <line x1="12" y1="8" x2="12" y2="16" />
+        <line x1="8" y1="12" x2="16" y2="12" />
+      </svg>
+      <h4>Start building your evaluator</h4>
+      <p>Add your first principle from the tree panel, then define requirements for each one.</p>
+    </div>
+  );
+}
+
 function RootDetail({ standard, onUpdateField, editable, isNew }) {
   const handleNameChange = (e) => {
     const name = e.target.value;
@@ -64,17 +78,7 @@ function RootDetail({ standard, onUpdateField, editable, isNew }) {
         </p>
       )}
 
-      {editable && (!standard.principles || standard.principles.length === 0) && (
-        <div className="detail-empty-state">
-          <svg width="36" height="36" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-            <circle cx="12" cy="12" r="10" />
-            <line x1="12" y1="8" x2="12" y2="16" />
-            <line x1="8" y1="12" x2="16" y2="12" />
-          </svg>
-          <h4>Start building your evaluator</h4>
-          <p>Add your first principle from the tree panel, then define requirements for each one.</p>
-        </div>
-      )}
+      {editable && (!standard.principles || standard.principles.length === 0) && <EmptyState />}
     </div>
   );
 }
