+++ tests/engine/test_adaptive_scaling_integration.py
@@ -17,7 +17,6 @@ def _counting_run_analysis(call_log):
         stream_file.write_text("")
         call_log.append(config.agent_id)
         if config.queue_path:
-            from quodeq.engine.file_queue import FileQueue
             queue = FileQueue(config.queue_path)
             queue.take(queue.remaining(), agent_id=config.agent_id)
         if config.jsonl_file:
